Revert "Reorganize project (#18)"
This reverts commit 7320afea4904fe89ef28efe7428c15c50829d808.
