+++ examples/example2/main.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/krisukox/google-flights-api"
+	"github.com/krisukox/google-flights-api/flights"
 	"golang.org/x/text/currency"
 	"golang.org/x/text/language"
 )
