+++ flights/session.go
